requirements.txt
awscli==1.16.230
boto==2.48.0
boto3==1.9.220
botocore==1.12.220
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.4
Flask-Login==0.4.1
Flask-WTF==0.14.2
html5lib==1.0.1
hurry.filesize==0.9
Jinja2==2.10
pymongo==3.9.0
pyodbc==4.0.23
Werkzeug==0.14.1
WTForms==2.2.1

